create shopper api wrapper class for client
